Better enforce Zig version so people don't run into random roadblocks (#22)
Unfortunately, if you use an older version of Zig, there are other API differences in the build system that will pop up instead of our nice error message. But hopefully we can bring this pattern with us into the future. Also would be good to define a maximum version at some point.