Debug fishy gradients from `SoftMax` (#20)
Fix `SoftMax` activation function derivative calculations

Previously, the estimated numerical cost gradients for the weights/biases didn't match our actual calculated (analytical) cost gradient during backpropagation which indicated that our backpropagation was wrong somewhere. Now with this PR, the estimated numerical cost gradient and the actual (analytical) cost gradient matches thanks to the proper calculations being made (read on).

Essentially, the way I was calculating the derivative of the `SoftMax` function was flawed. I was only calculating the diagonals of the Jacobian Matrix where `k = i` and multiplying by cost vector.

Since `SoftMax` is the only activation (out of the activation functions we currently support) that uses multiple inputs to produce the output value. This means that the derivative of the activation function is more complicated (more than just a single value) and involves a whole Jacobian matrix that we weren't taking into account before.

This PR introduces a `gradient` function for activation functions that returns a list of derivatives (a row from that Jacobian matrix) to dot product with the cost derivative vector. Before we we're only multiplying by a single derivative from that Jacobian matrix (the diagonals where `k = i`) which gives flawed results but surprisingly still allowed the network to converge before.

For single-input activation functions, we continue to use the `derivative` function which is an efficient shortcut we can do in those cases (see the developer notes for more details on why we can do this).

⭐️ Special shoutout to Hans Musgrave ([@hmusgrave](https://github.com/hmusgrave)) for the immense amount of help to get my head around these concepts as I got stuck through this process. ⭐️

The developer notes hold more context and information on all of this.